Sort object IDs numerically
From https://bthome.io/format/ -

"Object ids have to be applied in numerical order (from low to high) in
your advertisement. This will make sure that if you have a device
(sensor) that is broadcasting a new measurement type that is added in a
new (minor) BTHome update, while your BTHome receiver isn't updated yet
to the same version, it will still be able to receive the older
supported measurement types. A BTHome receiver will stop parsing object
ids as soon as it finds an object id that isn't supported."